+++ android/app/src/main/kotlin/com/repforge/repforge/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.repforge.repforge
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()